requirements.txt
asn1crypto==0.24.0
cffi==1.12.2
Click==7.0
cryptography==2.6.1
Flask==1.0.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
psycopg2==2.7.7
pycparser==2.19
pyOpenSSL==19.0.0
python-dateutil==2.8.0
PyYAML==5.1
six==1.12.0
Werkzeug==0.15.1
